feat: move abi to separate crate
