+++ crates/l1/src/lib.rs
@@ -1,0 +1,4 @@
+//! All L1 related primitives.
+
+/// L1 contracts Abi.
+pub mod abi;
