+++ crates/watcher/Cargo.toml
@@ -18,11 +18,12 @@ alloy-network.workspace = true
 alloy-primitives.workspace = true
 alloy-provider.workspace = true
 alloy-rpc-types-eth.workspace = true
-alloy-sol-types = "0.8.20"
+alloy-sol-types.workspace = true
 alloy-transport.workspace = true
 
 # rollup
 rollup-node-primitives.workspace = true
+scroll-l1.workspace = true
 
 # scroll
 scroll-alloy-consensus.workspace = true
@@ -43,6 +44,7 @@ alloy-rpc-types-eth = { workspace = true, features = ["arbitrary"] }
 arbitrary.workspace = true
 eyre.workspace = true
 rand = "0.9"
+scroll-l1 = { workspace = true, features = ["test-utils"] }
 tracing-subscriber = "0.3"
 
 [features]
@@ -53,6 +55,5 @@ arbitrary = [
     "alloy-eips/arbitrary",
     "alloy-primitives/arbitrary",
     "alloy-rpc-types-eth/arbitrary",
-    "alloy-sol-types/arbitrary",
     "scroll-alloy-consensus/arbitrary",
 ]
