+++ crates/watcher/src/constants.rs
@@ -1,9 +1,9 @@
-use crate::contract::{CommitBatch, QueueTransaction};
 use std::sync::LazyLock;
 
 use alloy_primitives::{address, Address};
 use alloy_rpc_types_eth::Filter;
 use alloy_sol_types::SolEvent;
+use scroll_l1::abi::logs::{CommitBatch, QueueTransaction};
 
 /// The address of the Scroll Rollup contract on the L1.
 pub const ROLLUP_CONTRACT_ADDRESS: Address = address!("0xa13BAF47339d63B743e7Da8741db5456DAc1E556");
